Move feature app queries to its own folder
