+++ src/controllers/Admin-Web-App/appQueries/projects.js
@@ -14,10 +14,6 @@ const {
     QueryMostRecentlyUpdatedVariable, 
 } = require('../../../helpers/common-queries/variable-queries')
 const removeSensitiveProjectData = require('../../../helpers/data-cleaners-react-app/projects')
-const { 
-    removeSensitiveFeatureData, 
-    returnFeatureNames,  
-} = require('../../../helpers/data-cleaners-react-app/features')
 const removeSensitiveVariableData = require('../../../helpers/data-cleaners-react-app/variables')
 const cleanHomePageData = require('../../../helpers/data-cleaners-react-app/home')
 const formatApiKeyData = require('../../../helpers/data-cleaners-react-app/apiKeys')
@@ -37,19 +33,6 @@ async function getUserProjects (req, res) {
         ]) 
         let cleanData = removeSensitiveProjectData(userProjects, userFeatures, userVariables)
         return res.json(cleanData)
-    } catch (error) {
-        console.error(error)
-        res.sendStatus(500)
-    }
-}
-
-
-async function getUserFeatures (req, res) {
-    try {
-        let { user } = req
-        let userFeatures = await QueryFeaturesByUser(user)
-        let cleaned_data = removeSensitiveFeatureData(userFeatures)
-        res.json(cleaned_data)
     } catch (error) {
         console.error(error)
         res.sendStatus(500)
@@ -118,25 +101,9 @@ async function getHomePageData (req, res) {
     }
 }
 
-async function GetFeaturesByProjectName (req, res) {
-    try {
-        let { user } = req 
-        let projectName = req.query.project_name
-        let features = await QueryFeaturesByProject(projectName, user)
-        let getNames = returnFeatureNames(features)
-        res.json(getNames)
-    } catch (error) {
-        console.error(error)
-        res.sendStatus(500)
-    }
-}
-
-
 module.exports = { 
     getUserProjects, 
-    getUserFeatures, 
     getVariables, 
     getApiKeys, 
     getHomePageData, 
-    GetFeaturesByProjectName,
 }
